requirements.txt
html_table_parser_python3~=0.3.1
requests~=2.28.2
